# BackEnd/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TsunamiIT Manager Project API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/project/{idOwner}": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Returns All Project By Owner ID",
                "operationId": "a5a7424f8ae7d05325dd8949679a168e",
                "parameters": [
                    {
                        "name": "idOwner",
                        "in": "path",
                        "description": "Owner ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation !"
                    },
                    "404": {
                        "description": "Project not found "
                    },
                    "500": {
                        "description": "Server error "
                    }
                }
            }
        },
        "/api/project/{idProject}": {
            "put": {
                "tags": [
                    "Project"
                ],
                "summary": "Update Project name, description and githubRepo",
                "operationId": "7cd410edebd66ff0e6411e16dc2a37af",
                "parameters": [
                    {
                        "name": "idProject",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Platforme"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Platforme E-learning"
                                    },
                                    "githubRepo": {
                                        "type": "string",
                                        "example": "github.com/..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Project"
                ],
                "summary": "Remove a Project If Project  isn't currently being worked on by a SCRUM team",
                "operationId": "aac8ce98991aa64ee9e5d55440850c30",
                "parameters": [
                    {
                        "name": "idProject",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            }
        },
        "/api/project": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Add Project",
                "operationId": "29c36ea882b930f669023c405ec68290",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name,type,githubRepo"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "IT"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "LMS"
                                    },
                                    "githubRepo": {
                                        "type": "string",
                                        "example": "MohamedNabli004"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/projects": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Returns all projects",
                "operationId": "6c7b9ebd190e43515069606a0a422b0e",
                "responses": {
                    "200": {
                        "description": "Successful operation !"
                    },
                    "204": {
                        "description": "Project has no versions yet !"
                    },
                    "404": {
                        "description": "No project found "
                    },
                    "500": {
                        "description": "Server error "
                    }
                }
            }
        },
        "/api/collaborator/{idCollaborator}/projects": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Get projects worked on by collaborator",
                "operationId": "9abdc84d70192f6de4e5f056b75b93ca",
                "parameters": [
                    {
                        "name": "idCollaborator",
                        "in": "path",
                        "description": "Collaborator ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "No affected team or project !"
                    }
                }
            }
        },
...
    }
}